Temporary Bot with Limitations
This version of the bot introduces about half of the desired features,
with many limitations on how the bot can be used. This is intended as a
stop-gap to allow games to continue, and to provide a base
implementation to tweak and improve over time to get back to feature
parity.